Move memory storage into separate file.
